refactor: move handle logic to ImportButton
