+++ src/app/App.tsx
@@ -1,15 +1,12 @@
-import apiFetch from '@wordpress/api-fetch';
 import { useState } from '@wordpress/element';
-import type { MediaFrame } from '@wordpress/media-utils';
 
 import Purge from '../components/Purge';
 import Notice from '../components/Notice';
-import ProgressBar from '../components/ProgressBar';
-import ImportButton from '../components/ImportButton';
 import TableName from '../components/TableName';
+import ProgressBar from '../components/ProgressBar';
 import TableColumns from '../components/TableColumns';
+import ImportButton from '../components/ImportButton';
 
-import { getModalParams } from '../utils';
 import '../styles/app.scss';
 
 interface SQLProps {
@@ -39,107 +36,14 @@ const App = (): JSX.Element => {
     }
   );
 
-  /**
-   * Handle Upload.
-   *
-   * This function is responsible for opening the
-   * WP media modal to enable user select.
-   *
-   * @since 1.0.0
-   *
-   * @returns {void}
-   */
-  const handleUpload = (): void => {
-    const wpMediaModal = wp.media( getModalParams() );
-    wpMediaModal.on( 'select', () => handleSelect( wpMediaModal ) ).open();
-  };
-
-  /**
-   * Handle Selection.
-   *
-   * This function is responsible for handling a
-   * selection made by the user.
-   *
-   * @since 1.0.0
-   *
-   * @param {MediaFrame} wpMediaModal WP Media Modal.
-   * @returns Promise<void>
-   */
-  const handleSelect = async (wpMediaModal: MediaFrame): Promise<void> => {
-    const args = wpMediaModal.state().get( 'selection' ).first().toJSON();
-
-    // Reset.
-    setSqlNotice( '' );
-    setParsedSQL(
-      {
-        tableName: '',
-        tableColumns: [],
-        tableRows: [],
-      }
-    );
-    setIsLoading( true );
-
-    // Parse SQL.
-    try {
-      setParsedSQL(
-        await apiFetch(
-          {
-            path: '/sql-to-cpt/v1/parse',
-            method: 'POST',
-            data: {
-              ...args
-            },
-          }
-        )
-      );
-      setIsLoading( false );
-    } catch ( { message } ) {
-      setIsLoading( false );
-      setSqlNotice( message );
-    }
-  };
-
-  /**
-   * Handle Import.
-   *
-   * This function is responsible for handling the
-   * import made by the user.
-   *
-   * @since 1.1.0
-   *
-   * @returns Promise<void>
-   */
-  const handleImport = async (): Promise<void> => {
-    setIsLoading( true );
-
-    // Import SQL.
-    try {
-      const url = await apiFetch(
-        {
-          path: '/sql-to-cpt/v1/import',
-          method: 'POST',
-          data: {
-            ...parsedSQL
-          },
-        }
-      );
-      if ( url ) {
-        window.location.href = `${url}`
-      }
-      setIsLoading( false );
-    } catch ( { message } ) {
-      setIsLoading( false );
-      setSqlNotice( message );
-    }
-  }
-
   return (
     <main>
       <section>
         <ImportButton
           parsedSQL={ parsedSQL }
-          handleImport={ handleImport }
-          handleUpload={ handleUpload }
+          setIsLoading={ setIsLoading }
+          setSqlNotice={ setSqlNotice }
+          setParsedSQL={ setParsedSQL }
         />
         <Purge
           setIsLoading={ setIsLoading }
